+++ agent/src/ebpf/user/ctrl_tracer.c
@@ -472,6 +472,10 @@ static int socktrace_do_cmd(struct df_bpf_obj *obj, df_bpf_cmd_t cmd,
 		       sk_trace_params->kern_trace_map_max);
 		printf("kern_trace_map_used:\t%u\n",
 		       sk_trace_params->kern_trace_map_used);
+		printf("proc_exec_event_count:\t%lu\n",
+		       sk_trace_params->proc_exec_event_count);
+		printf("proc_exit_event_count:\t%lu\n",
+		       sk_trace_params->proc_exit_event_count);
 		printf("datadump_enable:\t%s\n",
 		       sk_trace_params->datadump_enable ? "true" : "false");
 		printf("datadump_pid:\t%d\n", sk_trace_params->datadump_pid);
